requirements.txt
wheel==0.24.0

pytest
mock
coverage

click==3.3
invoke
pyaml
frigg-common
clint
